refactor: 정적 리소스 찾는 로직 이동
